py/requirements.txt
opencv-python
mediapipe
pyautogui
websocket-client
numpy
